Extract EditCommandDescriptor and standardize EditCommand
