+++ src/main/java/seedu/address/logic/commands/update/AddContactToTodoCommand.java
@@ -111,14 +111,14 @@ public class AddContactToTodoCommand extends EditCommand<Todo> {
             return false;
         }
 
-        return index.equals(otherCommand.index)
+        return targetIndex.equals(otherCommand.targetIndex)
                 && contactIndices.equals(otherCommand.contactIndices);
     }
 
     @Override
     public String toString() {
         return new ToStringBuilder(this)
-                .add("index", index)
+                .add("index", targetIndex)
                 .add("contactIndicies", contactIndices)
                 .toString();
     }
