+++ src/main/java/seedu/address/logic/commands/update/MarkTodoAsDoneCommand.java
@@ -78,13 +78,13 @@ public class MarkTodoAsDoneCommand extends EditCommand<Todo> {
             return false;
         }
 
-        return index.equals(otherCommand.index);
+        return targetIndex.equals(otherCommand.targetIndex);
     }
 
     @Override
     public String toString() {
         return new ToStringBuilder(this)
-                .add("index", index)
+                .add("index", targetIndex)
                 .toString();
     }
 }
