+++ src/main/java/seedu/address/logic/commands/update/MarkTodoAsNotDoneCommand.java
@@ -79,13 +79,13 @@ public class MarkTodoAsNotDoneCommand extends EditCommand<Todo> {
             return false;
         }
 
-        return index.equals(otherCommand.index);
+        return targetIndex.equals(otherCommand.targetIndex);
     }
 
     @Override
     public String toString() {
         return new ToStringBuilder(this)
-                .add("index", index)
+                .add("index", targetIndex)
                 .toString();
     }
 }
