+++ src/test/java/seedu/address/logic/commands/update/EditContactDescriptorTest.java
@@ -11,7 +11,6 @@ import static seedu.address.logic.commands.CommandTestUtil.VALID_TAG_HUSBAND;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.update.EditContactCommand.EditContactDescriptor;
 import seedu.address.testutil.EditContactDescriptorBuilder;
 
 public class EditContactDescriptorTest {
