+++ src/test/java/seedu/address/logic/parser/contact/EditContactCommandParserTest.java
@@ -27,7 +27,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.update.EditContactCommand;
-import seedu.address.logic.commands.update.EditContactCommand.EditContactDescriptor;
+import seedu.address.logic.commands.update.EditContactDescriptor;
 import seedu.address.model.contact.Email;
 import seedu.address.model.contact.Name;
 import seedu.address.model.contact.Tag;
