+++ src/test/java/seedu/address/testutil/ContactUtil.java
@@ -11,7 +11,7 @@ import static seedu.address.logic.parser.contact.ContactCliSyntax.PREFIX_CONTACT
 import java.util.Set;
 
 import seedu.address.logic.commands.create.AddContactCommand;
-import seedu.address.logic.commands.update.EditContactCommand.EditContactDescriptor;
+import seedu.address.logic.commands.update.EditContactDescriptor;
 import seedu.address.model.contact.Contact;
 import seedu.address.model.contact.Tag;
